refactor vue component folder structure
