+++ resources/js/components/service-types/ServiceTypeShow.vue
